requirements.txt
aiohttp==3.6.2
alembic==1.4.2
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
typing-extensions==3.7.4.3
yarl==1.5.1
